pyproject.toml
[project]
name = "agentic-app-quickstart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arize-phoenix==11.24.0",
    "openai>=1.99.1",
    "openai-agents>=0.2.4",
    "openinference-instrumentation-openai-agents>=1.1.1",
    "polars>=1.32.3",
    "rich>=14.1.0",
    "streamlit>=1.48.1",
    "tweepy>=4.16.0",
    "pandas>=2.0.0",
    "gradio>=4.0.0",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "marimo>=0.2.0",
    "opentelemetry-exporter-otlp>=1.36.0",
    "openinference-semantic-conventions>=0.1.21",
]

[dependency-groups]
dev = [
    "pyclean>=3.1.0",
    "pre-commit>=4.0.0",
    "pytest>=8.4.1",
    "ruff>=0.12.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agentic_app_quickstart"]
